refactored code for accessing main
