+++ core/interpreter.py
@@ -1,4 +1,6 @@
 from core.parser.nodes import Token, TokenType
+from .lexer.lexicalAnalyzer import Lexer
+from .parser.syntaxAnalyzer import Parser
 
 
 class Interpreter:
@@ -7,3 +9,20 @@ class Interpreter:
 
     def execute(self):
         return self.__root_node.get_node_value()
+
+
+def main(input_string: str):
+    lexer = Lexer(input_string)
+
+    tokens = lexer.get_tokens()
+    # result = lexer.get_tokens()
+
+    parser = Parser(tokens)
+
+    root_node = parser.parse()
+
+    interpreter = Interpreter(root_node)
+
+    result = interpreter.execute()
+
+    return result
